+++ src/Utility.h
@@ -5,4 +5,4 @@ void printError(std::string msg, int errcode, long timestamp);
 void printGraph(std::string title, int XDist, int YDist, std::string dataFilePath,int width, int height);
 double* seq(double start, double end, int size);
 bool fexists(std::string filename);
-void printSet(std::set<std::pair<double,double> > *pairs);
+void printSet(std::set<std::pair<int,int> > *pairs);
